miscellaneous leash32.dll corrections
do not build leash32.dll with MessageBox() calls.  Doing
so produces a library that on error must display a dialog
box to the end user.  If this library is called from a
service (via gssapi32.dll) the service will block forever
while awaiting the dialog box to clear.

LeashKRB5GetTickets() treats krbv5Context as an in/out variable.
If the caller does not provide a krb5_context, one will be allocated.
It is up to the caller to ensure that the context is eventually freed.
A context can be returned even if the function returns an error.
Make sure that 'ctx' and 'cache' are properly initialized so that
it is possible to tell the difference.  Do not free the context if
it was locally allocated.

In acquire_tkt_no_princ() do not set the KRB5CCNAME environment
variable if 'ccname' is an empty string.

Combination of 2 patches by Jeffrey Altman