Add new public header gssapi_alloc.h
Contains allocator methods for use with mechanisms and mechglues for
allocations that must be made in one module but freed in another.
On windows, an allocation made in one module cannot safely be freed in
another using the usual c runtime malloc/free; runtime dll mismatch will
cause heap corruption in that case.  But it is safe to instead directly
use HeapAlloc()/HeapFree() specifying the default process heap.

Signed-off-by: Kevin Wasserman <kevin.wasserman@painless-security.com>